Implement collision detection with pillars
Whew, that was trickier than I thought!

Even with this, there is still a slight bug that lurks. Sometimes,
when munging the arrow keys next to a wall, the sprite can become
stuck. It's possible to become unstuck if you change direction from
vertical to horizontal, in a specific way.

Also, this commit renames 'compute_coords' to 'compute_pixel_coords',
and introduces a new function 'compute_grid_coords'. I thought I was
going to need the latter at some point, but was wrong. Neverthess, it
looks like it could be useful eventually, so I'm leaving it for now.