Redefine player sprite as a MovingThing object
The plan is to also define crawlers as members of this
class. Hopefully it'll soon be easier to find ways of fixing the
crawler's animation and motion.

I also defined 'get_all' as a convenient "map" style operation for
deriving animation frames (which are of type 'pygame.Surface') from
TileDef objects.