Define player 'update' in terms of collision_type dict
The collision type dict works as follows. An entry of

"block": obstacle_group

means that any sprite in 'obstacle_group' is a blocking obstacle for
me; that is, I can't pass through it.

Let's use a hypothetical example. An entry of

"damage": crawler_group

would mean that the receiving sprite would receive damage upon
encountering a sprite in 'crawler_group'.