Eliminate CollisionType.DO_DAMAGE
That is, only keep 'CollisionType.TAKE_DAMAGE'. This is easier to work
with, because there doesn't need to be an associated 'update' method
with sprites listed here. On the other hand, DO_DAMAGE refers to the
sprite calling the update method, doing damage to another sprite.

However, this is a problem in the case of sprites like the one used to
display the sword, which is created extemporaneously within the Player
class when needed.